1st number

+++ operators.py
@@ -1,4 +1,4 @@
-x = float(input("enter the first number:"))
-y = float(input("enter the second number:"))
+x = int(input("enter the first number:"))
+y = int(input("enter the second number:"))
 sum = x + y
 print("the sum is:", sum)
